feat: idempotency key support for /api/loan/repay
- Extract in-memory idempotency cache to src/middleware/idempotency.ts
- 24h TTL, auto-eviction on access
- index.ts imports getIdempotencyEntry/setIdempotencyEntry from middleware
- Missing Idempotency-Key returns 400
- Duplicate key within 24h returns cached response with X-Idempotent-Replayed: true
- Document X-Idempotent-Replayed response header in openapi.json

Closes #5
